Build and pack a WASM rewriter
