+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 edition = "2021"
-name = "datadog_native-iast-rewriter"
-version = "0.0.1"
+name = "datadog-native-iast-rewriter"
+version = "0.1.0-beta.3"
 
 [lib]
 crate-type = ["cdylib"]
@@ -17,6 +17,7 @@ swc_ecma_parser = "0.111.0"
 anyhow = "1.0.56"
 fastrand = "1.8.0"
 base64 = "0.13.0"
+wasm-bindgen = "0.2"
 
 [dev-dependencies]
 tempfile = "3.3.0"
@@ -28,3 +29,7 @@ napi-build = "1.2.1"
 
 [profile.release]
 lto = true
+
+[features]
+default = []
+wasm = []
