+++ LICENSE-3rdparty.csv
@@ -9,6 +9,7 @@ require,anyhow,MIT OR Apache-2.0,Copyright (c) 2018 David Tolnay
 require,fastrand,Apache-2.0 OR MIT, Copyright 2020 Stjepan Glavina
 require,base64,MIT and Apache-2.0,Copyright (c) 2015 Alice Maz
 require,node-gyp-build,MIT,Copyright 2017 Mathias Buus
+require,wasm-bindgen,Apache-2.0 OR MIT,Copyright (c) 2014 Alex Crichton
 dev,tempfile,MIT OR Apache-2.0,Copyright (c) 2015 Steven Allen
 dev,spectral,Apache-2.0,Copyright (c) cfrancia
 dev,ctor,Apache-2.0 OR MIT,Copyright (c) 2018 Matt Mastracci
