+++ package.json
@@ -78,6 +78,8 @@
     "build:windows": "napi build --target x86_64-pc-windows-msvc --platform --release",
     "build:darwin": "napi build --target x86_64-apple-darwin --platform --release",
     "build:darwinArm": "napi build --target aarch64-apple-darwin --platform --release",
+    "build:wasm": "wasm-pack build --features \"wasm\"",
+    "pack:wasm": "npm run build:wasm && wasm-pack pack",
     "buildAndTest:debug": "npm run build && npm run test:junit",
     "buildAndTest:linux": "npm run build:linux && npm run test:junit",
     "buildAndTest:alpine": "npm run build:alpine && npm run test:junit",
