+++ src/lib.rs
@@ -10,42 +10,12 @@ mod visitor;
 #[cfg(test)]
 mod tests;
 
+#[cfg(feature = "wasm")]
+mod lib_wasm;
+
 #[macro_use]
+#[cfg(not(feature = "wasm"))]
 extern crate napi_derive;
-extern crate base64;
 
-use crate::rewriter::{print_js, rewrite_js};
-use napi::{Error, Status};
-
-#[napi(object)]
-#[derive(Debug)]
-pub struct RewriterConfig {
-    pub chain_source_map: Option<bool>,
-    pub comments: Option<bool>,
-}
-
-#[napi]
-pub struct Rewriter {
-    config: RewriterConfig,
-}
-
-#[napi]
-impl Rewriter {
-    #[napi(constructor)]
-    pub fn new(config: Option<RewriterConfig>) -> Self {
-        let rewriter_config: RewriterConfig = config.unwrap_or(RewriterConfig {
-            chain_source_map: Some(false),
-            comments: Some(false),
-        });
-        Self {
-            config: rewriter_config,
-        }
-    }
-
-    #[napi]
-    pub fn rewrite(&self, code: String, file: String) -> napi::Result<String> {
-        rewrite_js(code, file, self.config.comments.unwrap_or(false))
-            .map(|result| print_js(result, self.config.chain_source_map.unwrap_or(false)))
-            .map_err(|e| Error::new(Status::Unknown, format!("{}", e)))
-    }
-}
+#[cfg(not(feature = "wasm"))]
+mod lib_napi;
